+++ packages/auth/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @emdash-cms/auth
 
+## 0.12.0
+
+## 0.11.1
+
 ## 0.11.0
 
 ### Patch Changes