+++ packages/plugins/embeds/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @emdash-cms/plugin-embeds
 
+## 0.1.13
+
+### Patch Changes
+
+- Updated dependencies [[`35791ff`](https://github.com/emdash-cms/emdash/commit/35791ff9f68c10c6d3ff15ee0ab407baef09c2aa), [`7b45cba`](https://github.com/emdash-cms/emdash/commit/7b45cba66143c3a75bbd880abff85303c1fd6072)]:
+  - emdash@0.12.0
+  - @emdash-cms/blocks@0.12.0
+
+## 0.1.12
+
+### Patch Changes
+
+- Updated dependencies [[`dc44989`](https://github.com/emdash-cms/emdash/commit/dc44989b263164625039525aa84e4a562f0a879f)]:
+  - emdash@0.11.1
+  - @emdash-cms/blocks@0.11.1
+
 ## 0.1.11
 
 ### Patch Changes